requirements.txt
Django==2.2.1
django-taggit==1.1.0
psycopg2==2.8.2
pytz==2019.1
sqlparse==0.3.0
